+++ lib/src/models/linked_showcase_data.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/widgets.dart';
+
+/// This model is used to move linked showcase overlay data to parent
+/// showcase to crop linked showcase rect
+class LinkedShowcaseDataModel {
+  final Rect rect;
+  final EdgeInsets overlayPadding;
+  final BorderRadius? radius;
+  final bool isCircle;
+
+  const LinkedShowcaseDataModel({
+    required this.rect,
+    required this.radius,
+    required this.overlayPadding,
+    required this.isCircle,
+  });
+}